requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
langchain>=0.2.0
langgraph>=0.2.0
langchain-ollama>=0.1.0
anthropic>=0.39.0
rich>=13.7.0
